pyproject.toml
[project]
name = "atb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "python-telegram-bot[job-queue]>=21.9",
    "pytz>=2024.2",
]

[dependency-groups]
dev = [
    "ruff>=0.8.4",
]
